+++ libs/components/src/lib/modules/overlay/overlay.component.ts
@@ -14,7 +14,7 @@ import {
 } from '@angular/core';
 import { PrizmDestroyService } from '@prizm-ui/helpers';
 import { PrizmThemeService } from '@prizm-ui/theme';
-import { Observable } from 'rxjs';
+import { Observable, timer } from 'rxjs';
 import { startWith, takeUntil, tap } from 'rxjs/operators';
 import { PrizmOverlayConfig, PrizmOverlayContent, PrizmOverlayContentType, PrizmOverlayId } from './models';
 import { PrizmOverlayAbstractPosition } from './position/position';
@@ -114,15 +114,24 @@ export class PrizmOverlayComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   private setPos(): void {
-    const positions = this.position.getPositions(this.wrapperEl);
-    const { extra, ...coords } = positions;
-    if (this.extra !== extra) {
-      this.extra = extra;
-      this.cd.detectChanges();
-    }
-    Object.assign(coords, { visibility: 'visible', opacity: '1' });
-    this.wrapperEl.style = objToCss(coords);
-    this.position.savePosition(positions);
-    EventBus.send(this.zid, 'z_posupdate');
+    timer(0)
+      .pipe(
+        tap(() => {
+          const positions = this.position.getPositions(this.wrapperEl);
+          const { extra, ...coords } = positions;
+          if (this.extra !== extra) {
+            this.extra = extra;
+            this.cd.detectChanges();
+          }
+          Object.assign(coords, { visibility: 'visible', opacity: '1' });
+          this.wrapperEl.style = objToCss(coords);
+          this.position.savePosition(positions);
+          EventBus.send(this.zid, 'z_posupdate');
+
+          this.cd.markForCheck();
+        }),
+        takeUntil(this.destroy$)
+      )
+      .subscribe();
   }
 }